+++ test/suite/AsplodeTest.php
@@ -3,7 +3,7 @@
 /*
  * This file is part of the Asplode package.
  *
- * Copyright © 2014 Erin Millard
+ * Copyright © 2016 Erin Millard
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -14,51 +14,49 @@ namespace Eloquent\Asplode;
 use ErrorException;
 use Icecave\Isolator\Isolator;
 use PHPUnit_Framework_TestCase;
-use Phake;
+use Phunky;
 
 class AsplodeTest extends PHPUnit_Framework_TestCase
 {
     protected function setUp()
     {
-        parent::setUp();
-
-        $this->isolator = Phake::mock('Icecave\Isolator\Isolator');
+        $this->isolator = Phunky::mock('Icecave\Isolator\Isolator');
     }
 
     public function testInstall()
     {
         Asplode::install($this->isolator);
 
-        Phake::verify($this->isolator)->set_error_handler($this->isInstanceOf(__NAMESPACE__ . '\ErrorHandler'));
-        Phake::verify($this->isolator)->register_shutdown_function($this->isInstanceOf(__NAMESPACE__ . '\FatalErrorHandler'));
+        Phunky::verify($this->isolator)->set_error_handler($this->isInstanceOf(__NAMESPACE__ . '\ErrorHandler'));
+        Phunky::verify($this->isolator)->register_shutdown_function($this->isInstanceOf(__NAMESPACE__ . '\FatalErrorHandler'));
     }
 
     public function testInstallErrorHandler()
     {
         Asplode::installErrorHandler($this->isolator);
 
-        Phake::verify($this->isolator)->set_error_handler($this->isInstanceOf(__NAMESPACE__ . '\ErrorHandler'));
-        Phake::verify($this->isolator, Phake::never())->register_shutdown_function(Phake::anyParameters());
+        Phunky::verify($this->isolator)->set_error_handler($this->isInstanceOf(__NAMESPACE__ . '\ErrorHandler'));
+        Phunky::verify($this->isolator, Phunky::never())->register_shutdown_function(Phunky::anyParameters());
     }
 
     public function testInstallFatalHandler()
     {
         Asplode::installFatalHandler($this->isolator);
 
-        Phake::verify($this->isolator)->register_shutdown_function($this->isInstanceOf(__NAMESPACE__ . '\FatalErrorHandler'));
-        Phake::verify($this->isolator, Phake::never())->set_error_handler(Phake::anyParameters());
+        Phunky::verify($this->isolator)->register_shutdown_function($this->isInstanceOf(__NAMESPACE__ . '\FatalErrorHandler'));
+        Phunky::verify($this->isolator, Phunky::never())->set_error_handler(Phunky::anyParameters());
     }
 
     public function testAssertCompatibleHandler()
     {
-        Phake::when($this->isolator)
-            ->trigger_error(Phake::anyParameters())
+        Phunky::when($this->isolator)
+            ->trigger_error(Phunky::anyParameters())
             ->thenGetReturnByLambda(function ($message, $errorType) {
                 throw new ErrorException($message, 0, $errorType, '/path/to/file', 111);
             });
         Asplode::assertCompatibleHandler($this->isolator);
 
-        Phake::verify($this->isolator)->trigger_error('Error handling is incorrectly configured.', E_USER_NOTICE);
+        Phunky::verify($this->isolator)->trigger_error('Error handling is incorrectly configured.', E_USER_NOTICE);
     }
 
     public function testAssertCompatibleHandlerFailure()