+++ test/suite/Exception/ErrorHandlingConfigurationExceptionTest.php
@@ -3,7 +3,7 @@
 /*
  * This file is part of the Asplode package.
  *
- * Copyright © 2014 Erin Millard
+ * Copyright © 2016 Erin Millard
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -18,7 +18,7 @@ class ErrorHandlingConfigurationExceptionTest extends PHPUnit_Framework_TestCase
 {
     public function testException()
     {
-        $previous = new Exception;
+        $previous = new Exception();
         $exception = new ErrorHandlingConfigurationException($previous);
 
         $this->assertSame('Error handling is incorrectly configured.', $exception->getMessage());