+++ test/suite/FunctionalTest.php
@@ -3,7 +3,7 @@
 /*
  * This file is part of the Asplode package.
  *
- * Copyright © 2014 Erin Millard
+ * Copyright © 2016 Erin Millard
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -16,10 +16,8 @@ class FunctionalTest extends PHPUnit_Framework_TestCase
 {
     protected function setUp()
     {
-        parent::setUp();
-
-        $this->errorHandlerStack = new ErrorHandlerStack;
-        $this->exceptionHandlerStack = new ExceptionHandlerStack;
+        $this->errorHandlerStack = new ErrorHandlerStack();
+        $this->exceptionHandlerStack = new ExceptionHandlerStack();
 
         $this->errorHandlers = $this->errorHandlerStack->clear();
         $this->exceptionHandlers = $this->exceptionHandlerStack->clear();
@@ -71,8 +69,10 @@ class FunctionalTest extends PHPUnit_Framework_TestCase
         Eloquent\Asplode\Asplode::install();
         $caught = false;
         try {
-            $callback(new stdClass);
+            $callback(new stdClass());
         } catch (ErrorException $e) {
+            $caught = true;
+        } catch (Error $e) {
             $caught = true;
         }
 
@@ -117,7 +117,7 @@ EOD;
         $output = implode(PHP_EOL, $output);
 
         $this->assertNotEquals(0, $exitCode);
-        $this->assertContains("Caught 'Call to undefined function foo()'", $output);
+        $this->assertContains('undefined function foo()', $output);
     }
 
     /**
@@ -150,5 +150,22 @@ EOD;
             "/Caught 'Allowed memory size of \d+ bytes exhausted \(tried to allocate \d+ bytes\)'/",
             $output
         );
+    }
+
+    public function testStackTrimming()
+    {
+        Eloquent\Asplode\Asplode::install();
+        $caught = false;
+        try {
+            ($line = __LINE__) && ($fp = fopen(uniqid(), 'r'));
+        } catch (ErrorException $e) {
+            $caught = true;
+        }
+        $file = __FILE__;
+
+        $this->assertTrue($caught);
+        $this->assertSame($file, $e->getFile());
+        $this->assertSame($line, $e->getLine());
+        $this->assertFalse(strpos($e->getTraceAsString(), 'Asplode'));
     }
 }